feat: Add stock_data_if.py for retrieving and storing stock data
This commit adds the stock_data_if.py file, which contains functions for retrieving and storing stock data. The file includes functions for getting stock lists from the Shanghai Stock Exchange (SSE) and Eastmoney, as well as functions for initializing and creating stock tables in a MySQL database. It also includes a function for retrieving daily stock data from Eastmoney and storing it in the corresponding stock table.

The functions in this file are essential for the data engine module and will be used to fetch and store stock data for further analysis.